backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==25.1.0
cffi==1.17.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
typing_extensions==4.12.2
yarl==1.18.3
